+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release
 
 * Saves and deletes now use a new database-agnostic locking mechanism using a sentinel row to properly serializes all tree modifications. That rules out page-tree-trashing by concurrent requests forever (a problem that was known to happen very rarely with [UnCMS's ancestor](history.md), though it has never been seen with UnCMS).
+* There is now a bulk file uploader for files in the [media app](media-app.md).
 * The test suite now passes with a MySQL/MariaDB database, so those databases can be considered to be supported.
 * Various [test factories](testing-your-project.md) have been retired and replaced with traits for `FileFactory`.
 * The in-browser image editor (inherited from UnCMS's ancestor) has been retired. It worked poorly on large files and it was a giant pile of vendored JS that hadn't been updated since 2019 and which I have little interest in keeping updated. Some day a simpler replacement giving the basic crop/rotate operations that this provided might be implemented in a vastly simpler way (see [#13](https://github.com/uncms-dev/uncms/issues/13), but that day is not today.